requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
scikit-learn
xgboost
prophet
statsmodels
python-docx
google-generativeai
openpyxl
plotly
